Before this commit, the program searched (in the first block, the first string) for a datetime substring and determined the position of the first char of that substring. So it was impossible to work with log files, where datetime substrings wasn't on the constant position in a string. For example: === 193.42.104.88 - - [28/Mar/2019:23:08:51 +0100] "GET /index.php?option=com_contact HTTP/1.1" 196.19.8.38 - - [28/Mar/2019:23:09:27 +0100] "GET /index.php?option=com_contact&view=contact&id=1 HTTP/1.1" === The datetime substrings (which are between []) have different start positions in strings.
To make it possible to work with such logs, the algo was modified.
Now the program doesn't determine a datetime substring position. Instead
it gets a string and searhes for datetime substring from left to right
till the end of a string. If a datetime substring was not found then
it takes the next string and searches there.

The functions get_first_dt_str_from_bz2_blk(), get_last_dt_str_from_bz2_blk()
were redesigned to have common algorithm as much as possible. As a
result, dt_substr_in_str() and get_str() functions were added.