Fix:
This is the case which was fixed:

A bz2 block ends with such 2 last strings:
===
159.203.109.175 - - [24/Feb/2019:08:09:55 +0100] "GET /templates/jp_hotel/css/template.css HTTP/1.1" 200 10004 "-" "Mozilla/5.0 (X11; Datanyze; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/65.0.3325.181 Safari/537.36" "-"
159.203.109.175 - - [24/Feb/2
===

A program takes last string (159.203.109.175 - - [24/Feb/2) and
searches for the datetime substring in it. As the is no such substring
(in full size) in the sting it should go to the previous string and
search a datetime substring there.

Before this fix the program didn't search datetime substring in
the previous string.